Replace game by round
Now a game is composed by selecting number of players + selecting characters + x number of rounds
