+++ test/markScoreAllowed.test.html
@@ -11,7 +11,7 @@
     <script type="module">
       import { assert } from "@jsenv/assert"
       import { CARD_ANIMALS, CARD_CHEST } from "src/cards/cards.js"
-      import { useMarkScoreAllowed } from "src/game.selectors.js"
+      import { useMarkScoreAllowed } from "src/round/round.selectors.js"
 
       // with 3 skulls
       {
