+++ test/rollDiceAllowed.test.html
@@ -10,7 +10,7 @@
   <body>
     <script type="module">
       import { assert } from "@jsenv/assert"
-      import { useRollDiceAllowed } from "src/game.selectors.js"
+      import { useRollDiceAllowed } from "src/round/round.selectors.js"
       import { CARD_COIN } from "src/cards/cards.js"
       import { createCoinFromDice } from "./test.material.js"
