+++ test/startNextRoundAllowed.test.html
@@ -10,7 +10,7 @@
   <body>
     <script type="module">
       import { assert } from "@jsenv/assert"
-      import { useStartNextRoundAllowed } from "src/game.selectors.js"
+      import { useStartNextRoundAllowed } from "src/round/round.selectors.js"
 
       // round ongoing because can roll dice
       {
